Se actualizo con dev y se modifico las vistas de condiciones
Aun me falta limpiar ese espacio en blanco restante en la vista de condiciones

+++ DP1WMS/src/main/java/com/dp1wms/controller/Guia/CrearGuiaController.java
@@ -66,7 +66,7 @@ public class CrearGuiaController implements FxmlController {
 
     private Envio envioEscogido;
 
-    private Double pesoTotal;
+    private float pesoTotal;
 
     private List<DetalleEnvio> listaDetalleEnvio;
 
@@ -89,10 +89,10 @@ public class CrearGuiaController implements FxmlController {
     public void registrarGuia(ActionEvent event){
         Guia guia = new Guia();
         List<DetalleGuia> listaDetalleGuia = new ArrayList<DetalleGuia>();
-        Double pesoTotal = 0.0;
+        Float pesoTotal = (float)0;
         for(DetalleEnvio detalleEnvio: this.listaDetalleEnvio){
             DetalleGuia detalleGuia = new DetalleGuia();
-            Double pesoProducto = detalleEnvio.getPeso();
+            Float pesoProducto = detalleEnvio.getPeso();
             if(pesoProducto != null){
                 detalleGuia.setPeso(pesoProducto);
                 pesoTotal += pesoProducto;
@@ -214,10 +214,10 @@ public class CrearGuiaController implements FxmlController {
         }
     }
 
-    private Double calcularPesoTotal() {
-        Double pesoTotal = 0.0;
+    private float calcularPesoTotal() {
+        float pesoTotal = 0;
         for (DetalleEnvio detalleEnvio : this.listaDetalleEnvio) {
-            Double pesoProducto = repositoryGuia.obtenerPesoProducto(detalleEnvio.getIdProducto());
+            Float pesoProducto = repositoryGuia.obtenerPesoProducto(detalleEnvio.getIdProducto());
             detalleEnvio.setPeso(pesoProducto);
             if (pesoProducto != null)
                 pesoTotal += pesoProducto;
@@ -238,7 +238,7 @@ public class CrearGuiaController implements FxmlController {
         this.listaDetalleEnvio = envioEscogido.getDetalleEnvio();
 
 
-        Double pesoTotal = calcularPesoTotal();
+        float pesoTotal = calcularPesoTotal();
         //pesoTotal = this.round(pesoTotal);
 
         this.txb_nombreCliente.setText(repositoryGuia.obtenerNombreCliente(this.envioEscogido.getIdCliente()));
